[update] change location query dto
