+++ packages/marketplace-service/src/app.module.ts
@@ -6,6 +6,8 @@ import configuration from '@app/shared/configuration';
 import { AppLoggerMiddleware } from '@app/shared/middleware/app-logger.middleware';
 import { NftModule } from './nfts/nfts.module';
 import { NftCollectionsModule } from './nft-collections/nft-collections.module';
+import { HistoryModule } from './histories/history.module';
+import { WalletModule } from './wallet/wallet.module';
 
 @Module({
   imports: [
@@ -13,6 +15,8 @@ import { NftCollectionsModule } from './nft-collections/nft-collections.module';
     UsersModule,
     MongooseModule.forRoot(configuration().db_path),
     NftModule,
+    HistoryModule,
+    WalletModule,
     NftCollectionsModule,
   ],
   controllers: [],
