+++ packages/marketplace-service/src/user/user.controller.ts
@@ -2,9 +2,10 @@ import { ApiTags } from '@nestjs/swagger';
 import { Controller, Get, Req } from '@nestjs/common';
 import { UserService } from './user.service';
 import { JWT, User } from '@app/shared/modules';
-import { iInfoToken } from '@app/shared/modules/dtos-query/auth.dto';
+
 import { BaseResult } from '@app/shared/types/base.result';
 import { UserDto } from '@app/shared/models';
+import { iInfoToken } from '@app/shared/modules/jwt/jwt.dto';
 
 @ApiTags('Users')
 @Controller('user')
