--- packages/shared/modules/dtos-query/filters/index.ts
@@ -1,3 +1,0 @@
-// Filter DTO Request
-export * from './nft.filter.dto';
-export * from './nft-collection.filter.dto';
