--- packages/shared/modules/dtos-query/index.ts
@@ -1,3 +1,0 @@
-// Response and Requeest DTO From Controller
-export * from './auth.dto';
-export * from './filters';
